Change "jwalterweatherman" to "jWalterWeatherman"
Since 'vgo' is case sensitive when building dependencies, I need the
jWalterWeatherman libraries to be imported with the right case.  There is a
change to fix this [1] but I do not have time to wait for it.  So I have forked
viper to implement the fix.  I will delete after the fix is merged.

[1] https://github.com/spf13/viper/pull/463/files